src/company_index/requirements.txt
scikit-learn
numpy
scipy
pandas
faiss-cpu
sentence-transformers
